Improves Riven details and fixes database update
Enhances Riven details display by adding a dedicated overview tab with more information and interactive controls.
This includes displaying stats based on selected rank and weapon, along with reroll, mastery rank, and polarity information.

Corrects a logic error that was preventing database name updates.

+++ src-tauri/src/commands/stock_riven.rs
@@ -270,7 +270,7 @@ pub async fn stock_riven_get_by_id(
         item.weapon_unique_name,
         item.mastery_rank,
         item.re_rolls,
-        item.sub_type.unwrap().rank.unwrap_or(0),
+        item.sub_type.unwrap().rank.unwrap_or(0) as i32,
         item.polarity.clone(),
         attributes
     )?);
